+++ plugin/core/main.py
@@ -10,9 +10,9 @@ except ImportError:
 import sublime_plugin
 import sublime
 
-from .url import filename_to_uri, uri_to_filename
+from .url import filename_to_uri
 from .protocol import (
-    Request, Notification, Range
+    Request, Notification
 )
 from .settings import (
     ClientConfig, settings, load_settings, unload_settings
@@ -32,6 +32,7 @@ from .documents import (
     initialize_document_sync, notify_did_open
 )
 from .diagnostics import handle_diagnostics, remove_diagnostics
+from .edit import apply_workspace_edit
 
 
 def plugin_loaded():
@@ -228,11 +229,6 @@ def start_server(server_binary_args, working_dir, env):
         exception_log("Failed to start server", err)
 
 
-def apply_workspace_edit(window, params):
-    edit = params.get('edit')
-    window.run_command('lsp_apply_workspace_edit', {'changes': edit})
-
-
 class LspRestartClientCommand(sublime_plugin.TextCommand):
     def is_enabled(self):
         return is_supported_view(self.view)
@@ -248,53 +244,3 @@ class LspStartClientCommand(sublime_plugin.TextCommand):
 
     def run(self, edit):
         start_active_view()
-
-
-class LspApplyWorkspaceEditCommand(sublime_plugin.WindowCommand):
-    def run(self, changes):
-        debug('workspace edit', changes)
-        if changes.get('changes'):
-            for uri, file_changes in changes.get('changes').items():
-                path = uri_to_filename(uri)
-                view = self.window.open_file(path)
-                if view:
-                    if view.is_loading():
-                        # TODO: wait for event instead.
-                        sublime.set_timeout_async(
-                            lambda: view.run_command('lsp_apply_document_edit', {'changes': file_changes}),
-                            500
-                        )
-                    else:
-                        view.run_command('lsp_apply_document_edit',
-                                         {'changes': file_changes})
-                else:
-                    debug('view not found to apply', path, file_changes)
-
-
-class LspApplyDocumentEditCommand(sublime_plugin.TextCommand):
-    def run(self, edit, changes):
-        regions = list(self.create_region(change) for change in changes)
-        replacements = list(change.get('newText') for change in changes)
-
-        self.view.add_regions('lsp_edit', regions, "source.python")
-
-        index = 0
-        # use regions from view as they are correctly updated after edits.
-        for newText in replacements:
-            region = self.view.get_regions('lsp_edit')[index]
-            self.apply_change(region, newText, edit)
-            index += 1
-
-        self.view.erase_regions('lsp_edit')
-
-    def create_region(self, change):
-        return Range.from_lsp(change['range']).to_region(self.view)
-
-    def apply_change(self, region, newText, edit):
-        if region.empty():
-            self.view.insert(edit, region.a, newText)
-        else:
-            if len(newText) > 0:
-                self.view.replace(edit, region, newText)
-            else:
-                self.view.erase(edit, region)
